+++ lib/src/utils/after_layout.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 /// Callback after build widget is rendered
 /// CREDIT
 /// https://pub.dev/packages/after_layout
+@Deprecated('Use afterBuildCreated() instead')
 mixin AfterLayoutMixin<T extends StatefulWidget> on State<T> {
   @override
   void initState() {